TOML: move toml lexer to `lexer` package
