+++ intellij-toml/src/main/grammars/TomlLexer.flex
@@ -1,4 +1,4 @@
-package org.toml.lang.parse;
+package org.toml.lang.lexer;
 import com.intellij.lexer.*;
 import com.intellij.psi.tree.IElementType;
 import static org.toml.lang.psi.TomlElementTypes.*;
